Move rotation_matrix_from_normal to geom_utils
